Added create/update sort to admin company page
